+++ convex/game.ts
@@ -190,7 +190,9 @@ export const createNewGame = mutation({
         .withIndex("byUserClerkId", (q) => q.eq("userClerkId", identity.subject))
         .collect();
       for (const staleGame of staleOngoingGames) {
-        await ctx.db.delete(staleGame._id);
+        if (staleGame.gameType === "singleplayer") {
+          await ctx.db.delete(staleGame._id);
+        }
       }
     }
 
@@ -605,6 +607,7 @@ export const checkGuess = mutation({
     if (existingOngoingGame) {
       await ctx.scheduler.runAfter(0, internal.continuegame.deleteStaleGames, {
         userClerkId: identity.subject,
+        gameType: game.gameType,
         matchingId: existingOngoingGame._id,
       });
       await ctx.db.patch(existingOngoingGame._id, {
@@ -623,6 +626,7 @@ export const checkGuess = mutation({
       });
       await ctx.scheduler.runAfter(0, internal.continuegame.deleteStaleGames, {
         userClerkId: identity.subject,
+        gameType: game.gameType,
         matchingId: newId,
       });
     }
